Make repo installable - according to ros convention, move headers to include/apriltags2/ - change headers in apriltags2_ros accordingly - add install rules for apriltags2_ros
