+++ apriltags2/include/apriltags2/common/doubles.h
