+++ apriltags2/include/apriltags2/common/doubles_floats_impl.h
