+++ apriltags2/include/apriltags2/common/floats.h
