+++ apriltags2/include/apriltags2/common/g2d.h
