+++ apriltags2/include/apriltags2/common/image_f32.h
