+++ apriltags2/include/apriltags2/common/image_types.h
