+++ apriltags2/include/apriltags2/common/image_u8x4.h
