+++ apriltags2/include/apriltags2/common/pjpeg.h
