+++ apriltags2/include/apriltags2/common/string_util.h
