+++ apriltags2/include/apriltags2/common/workerpool.h
